fix: refactor test utils structure
