+++ contracts/test/utils/constants/errorMessages.ts
@@ -1,0 +1,2 @@
+export const INITIALIZED_ERROR_MESSAGE = "Initializable: contract is not initializing";
+export const INITIALIZED_ALREADY_MESSAGE = "Initializable: contract is already initialized";
