+++ contracts/test/utils/constants/messageStatuses.ts
@@ -1,0 +1,7 @@
+export const INBOX_STATUS_UNKNOWN = 0;
+export const INBOX_STATUS_RECEIVED = 1;
+export const INBOX_STATUS_CLAIMED = 2;
+
+export const OUTBOX_STATUS_UNKNOWN = 0;
+export const OUTBOX_STATUS_SENT = 1;
+export const OUTBOX_STATUS_RECEIVED = 2;
