+++ contracts/test/utils/constants/time.ts
@@ -1,0 +1,12 @@
+export const ONE_SECOND = 1;
+export const ONE_MINUTE_IN_SECONDS = 60 * ONE_SECOND;
+export const ONE_HOUR_IN_SECONDS = 60 * ONE_MINUTE_IN_SECONDS;
+export const ONE_DAY_IN_SECONDS = 24 * ONE_HOUR_IN_SECONDS;
+export const ONE_MONTH_IN_SECONDS = 30 * ONE_DAY_IN_SECONDS;
+export const ONE_YEARS_IN_SECONDS = 365 * ONE_DAY_IN_SECONDS;
+
+export const INITIAL_MIGRATION_BLOCK = 0;
+export const GENESIS_L2_TIMESTAMP = 0;
+export const DEFAULT_LAST_FINALIZED_TIMESTAMP = 1683325137n;
+export const SIX_MONTHS_IN_SECONDS = ONE_YEARS_IN_SECONDS / 2;
+export const TEST_PUBLIC_VERIFIER_INDEX = 0;
